transx_web/pyproject.toml
[tool.poetry]
name = "transx_web"
version = "0.1.0"
description = ""
authors = ["Julio Faerman <356476+faermanj@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "transx_web"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
crispy-bootstrap5 = "^2024.2"
social-auth-app-django = "^5.4.1"
google-api-python-client = "^2.131.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
jwt = "^1.3.1"
attrs = "^23.2.0"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
django-environ = "^0.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
